add logout functionality
